+++ CookHouse-app/src/components/adminModule/EventsTable.jsx
@@ -39,15 +39,21 @@ const EventsTable = (refreshVar) => {
 
   const handleDeletion = async (eventId) => {
     try {
-      const response = await axios.delete(
-        `${EVENTS_API_END_POINT}/delete/${eventId}`,
-        {
-          withCredentials: true,
-        }
-      );
+      if (window.confirm("Are you sure you want to cancel this event?")) {
+        // User clicked OK, proceed with deletion
+        const response = await axios.delete(
+          `${EVENTS_API_END_POINT}/delete/${eventId}`,
+          {
+            withCredentials: true,
+          }
+        );
 
-      if (response.data.success) {
-        toast.success(response.data.message);
+        if (response.data.success) {
+          toast.success(response.data.message);
+        }
+      } else {
+        // User clicked Cancel
+        toast.error("Deletion cancelled.");
       }
     } catch (error) {
       toast.error(error.response.data.message);
